+++ src/pages/ListsPage/ListsPage.tsx
@@ -1,10 +1,33 @@
-import { ArmyEditor, Page } from "../../widgets";
+import { useNavigate } from "react-router-dom";
+import { ArmyPicker, Page, Title } from "../../widgets";
+import { ArmyDescriptor } from "../../data/armies";
+import { useArmyDescriptorsStore } from "../../state/armyDescriptors";
+import { prepareArmyData } from "../../logic/prepareArmyData";
 
 export function ListsPage() {
 
+    const navigate = useNavigate();
+
+    const armyDescriptorsStore = useArmyDescriptorsStore();
+
+    const handleArmyPick = (army: ArmyDescriptor) => {
+
+        navigate(`/list/${army.id}`);
+    };
+
+    const handleCreateNewClick = () => {
+
+        const army = prepareArmyData({ });
+        armyDescriptorsStore.store(army);
+        
+        navigate(`/list/${army.id}`);
+    };
+
     return (
         <Page>
-            <ArmyEditor/>
+            <Title text="Lists"/>
+            <button onClick={handleCreateNewClick}>Create new</button>
+            <ArmyPicker onPick={handleArmyPick}/>
         </Page>
     );
 };