+++ src/widgets/ListEditor/ListEditor.tsx
@@ -1,6 +1,0 @@
-export function ListEditor() {
-
-    return (
-        <div>List editor</div>
-    );
-}