+++ src/widgets/WrittenField/WrittenField.tsx
@@ -1,13 +1,24 @@
+import { ChangeEvent, MutableRefObject } from "react";
 import "./WrittenField.css";
 
 export interface WrittenFieldProps {
     name: string;
     list?: string;
+    valueRef?: MutableRefObject<string>;
+    onChange?: (newValue: string) => void;
     defaultValue?: string;
     placeholder?: string;
 };
 
-export function WrittenField({ name, placeholder, list, defaultValue } : WrittenFieldProps) {
+export function WrittenField({ name, placeholder, list, valueRef, onChange, defaultValue } : WrittenFieldProps) {
+
+    const handleChange = (event: ChangeEvent) => {
+
+        const newValue = String((event.target as HTMLInputElement).value);
+
+        if (valueRef) valueRef.current = newValue;
+        onChange?.(newValue)
+    };
 
     return (
         <input
@@ -15,8 +26,9 @@ export function WrittenField({ name, placeholder, list, defaultValue } : Written
             type="text"
             list={list}
             name={name}
+            onChange={handleChange}
             placeholder={placeholder}
-            defaultValue={defaultValue}
+            defaultValue={defaultValue || (valueRef ? valueRef.current : '')}
         />
     );
 };